feat: dry run

+++ src/FileProcessor.ts
@@ -2,7 +2,10 @@ import * as fs from 'node:fs'
 import type { ImportRewriter } from './ImportRewriter.js'
 
 export class FileProcessor {
-  constructor(private readonly importRewriter: ImportRewriter) {}
+  constructor(
+    private readonly importRewriter: ImportRewriter,
+    private readonly dryRun: boolean,
+  ) {}
 
   processFile(filePath: string): boolean {
     const fileContent = fs.readFileSync(filePath, 'utf8')
@@ -12,7 +15,7 @@ export class FileProcessor {
       fileContent,
     )
 
-    if (modified) {
+    if (modified && !this.dryRun) {
       fs.writeFileSync(filePath, content, 'utf8')
     }
